requirements.txt
numpy==1.26.4
onnxruntime==1.17.0
pillow==10.2.0
opencv-python==4.9.0
